Updated SQLite from version 2 -> 3. PHP min 5.4.
If you are currently using SQLite as the cache method. Please be sure to delete the catalog/includes/modules/ultimate_seo_urls5/cache_system/sqlite/usu_cache.db file and "cycle" your cache method once to recreate the usu_cache.db to the latest version 3. Yeah, 2 and 3 are not compatible.